move markers into parser, since it is a part of this phase
