+++ frb_codegen/src/parser/ty.rs
@@ -1,6 +1,6 @@
 use crate::ir::IrType::*;
 use crate::ir::*;
-use crate::markers;
+use crate::parser::markers;
 use crate::parser::{extract_comments, extract_metadata, type_to_string};
 use crate::parser::source_graph::{Enum, Struct};
 use std::collections::{HashMap, HashSet};
